Refactor Level 2 UI into button prefabs
Replace numerous hand-authored Level 2 UI entities with prefab-based button instances and reorganize the UI hierarchy. Renamed UI elements (pause, joystick, collect/deliver, hide, hearts) into prefab instances under a new 'buttons' root, updated transforms, textures and layers, and swapped two button payloads from the Tutorial scene to mainmenu. Added new prefabs (buttons.prefab and collect deliver button.prefab) and their descriptors. Also added camera bookmarks and prefab overrides in Level1, and updated the Level 2 scene descriptor hash.

+++ PAINEngine/src/PAINEngine/Systems/Render/sysRender.cpp
@@ -1546,7 +1546,7 @@ namespace PAIN {
 						struct WallMinimapCache {
 							const entt::registry* sourceRegistry = nullptr;
 							size_t wallEntityCount = 0;
-							size_t wallEntitySignature = 0;
+							uint64_t wallEntitySignature = 0;
 							bool built = false;
 							bool gpuDirty = false; // true when triangleVerticesXZ changed and needs GPU re-upload
 							std::vector<glm::vec2> triangleVerticesXZ;
@@ -1606,7 +1606,7 @@ namespace PAIN {
 							wallCache.wallEntityCount = wallEntities.size();
 							size_t signature = 1469598103934665603ull;
 							for (const entt::entity entity : wallEntities) {
-								signature ^= static_cast<size_t>(static_cast<uint32_t>(entity));
+								signature ^= static_cast<uint64_t>(entity);
 								signature *= 1099511628211ull;
 							}
 							wallCache.wallEntitySignature = signature;
